change raw cake copy strategy; add deepcopy ability
