+++ tests/test_bakery.py
@@ -2,7 +2,6 @@
 
 import sys
 import types
-from copy import copy
 from dataclasses import dataclass
 from typing import (
     Any,
@@ -19,17 +18,7 @@ from typing import (
 
 import pytest
 
-from bakery import (
-    Bakery,
-    Cake,
-    PieceOfCake,
-    bake,
-    cake_baking_method,
-    cake_ingredients,
-    cake_name,
-    is_baked,
-    unbake,
-)
+from bakery import Bakery, Cake, PieceOfCake, bake, cake_name, is_baked, unbake
 from . import aclosing
 
 
@@ -446,59 +435,6 @@ async def test_bakery_caches() -> None:
     await MyBakery.aclose()
 
     assert getattr(MyBakery, "__bakery_items__") == bakery_items1
-
-
-async def test_copy_cake() -> None:
-    """Test new recipe."""
-
-    class CPU:
-        """Settings."""
-
-        def __init__(self, core_num: int, manufacturer: str):
-            self.core_num: int = core_num
-            self.manufacturer: str = manufacturer
-
-    class MyPC(Bakery):
-        """My bakery."""
-
-        core_num: int = Cake(4)
-        manufacturer: str = Cake("Intel")
-        cpu: CPU = Cake(CPU, core_num=core_num, manufacturer=manufacturer)
-
-    with pytest.raises(ValueError):  # recipe is not baked
-        _ = copy(MyPC.cpu)
-
-    async with MyPC():
-        cpu_cake = MyPC.cpu
-        cpu_new_cake = copy(MyPC.cpu)
-
-        assert id(cpu_cake) != id(cpu_new_cake)
-        assert cake_name(cpu_cake) == cake_name(cpu_new_cake)
-        assert cake_baking_method(cpu_cake) == cake_baking_method(cpu_new_cake)
-        assert True == is_baked(cpu_cake) != is_baked(cpu_new_cake)
-        assert cake_ingredients(cpu_cake).recipe is cake_ingredients(cpu_new_cake).recipe
-        assert cake_ingredients(cpu_cake).recipe_args == cake_ingredients(cpu_new_cake).recipe_args
-        assert (
-            cake_ingredients(cpu_cake).recipe_kwargs
-            == cake_ingredients(cpu_new_cake).recipe_kwargs
-        )
-
-        cpu: CPU = cpu_cake()
-        with pytest.raises(ValueError):
-            _ = cpu_new_cake.core_num()
-
-        new_cpu: CPU
-        async with cpu_new_cake as new_cpu:
-            assert id(cpu) != id(new_cpu)
-            assert cpu.core_num == new_cpu.core_num
-            assert cpu.manufacturer == new_cpu.manufacturer
-
-        new_cpu = await bake(cpu_new_cake)
-        assert id(cpu) != id(new_cpu)
-        assert cpu.core_num == new_cpu.core_num
-        assert cpu.manufacturer == new_cpu.manufacturer
-
-        await unbake(cpu_new_cake)
 
 
 async def test_object_wrapped_is_object() -> None:
